+++ src/intersection.rs
@@ -4,8 +4,8 @@ use na::Vector3;
 use std::rc::Rc;
 
 pub struct Intersection {
-    pub point: Point3<f64>,
-    pub normal: Vector3<f64>,
-    pub t: f64,
+    pub point: Point3<f32>,
+    pub normal: Vector3<f32>,
+    pub t: f32,
     pub material: Rc<dyn Material>,
 }
