+++ src/math.rs
@@ -1,3 +1,3 @@
-pub type Vec3 = na::Vector3<f64>;
-pub type Point3 = na::Point3<f64>;
-pub type Matrix4 = na::Matrix4<f64>;
+pub type Vec3 = na::Vector3<f32>;
+pub type Point3 = na::Point3<f32>;
+pub type Matrix4 = na::Matrix4<f32>;
